pyproject.toml
[project]
name = "data-pipelines-with-apache-airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "airflow-movielens",
    "apache-airflow==2.9.1",
    "apache-airflow-providers-postgres==5.8.0",
    "click>=8.3.0",
    "flask>=2.2.5",
    "flask-httpauth>=4.8.0",
    "pandas>=2.3.3",
    "polars>=1.35.2",
    "pytest>=9.0.0",
    "pytest-docker-tools>=3.1.9",
    "pytest-mock>=3.15.1",
    "setuptools>=80.9.0",
]

[tool.uv.workspace]
members = [
    "src/airflow-movielens",
]

[tool.uv.sources]
airflow-movielens = { workspace = true }
